% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipn_scale_basic.R
\name{ipn_scale_color_basic}
\alias{ipn_scale_color_basic}
\title{IPN Color palette scale: color}
\usage{
ipn_scale_color_basic(fl = 6, name, labels)
}
\arguments{
\item{fl}{set Forschungslinie, default set to "6" which is IPN blue}

\item{name}{set scale name}

\item{labels}{set scale labels}
}
\description{
IPN Color palette scale: color
}
\examples{
ipn_scale_color_basic( fl = 2 name = c("example scale"), labels = c("A","B"))

}
