% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipn_plot_basic.R
\name{ipn_theme_basic}
\alias{ipn_theme_basic}
\title{IPN theme uses IPN font, colors, and larger text size}
\usage{
ipn_theme_basic(fl = 6, size = 16, position = "right", grid = 0)
}
\arguments{
\item{fl}{set Forschungslinie, default set to "6" which is IPN blue}

\item{size}{set font size, default is 16}

\item{position}{set position for legend, default is "right"}

\item{grid}{show grid, default is 0 for no grid, set to 1 for grid}
}
\description{
IPN theme uses IPN font, colors, and larger text size
}
\examples{
ipn_style_basic(fl = 6, size = 16, position = "right", grid = 0)

}
