% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipn_scale_basic.R
\name{scale_color_ipn}
\alias{scale_color_ipn}
\title{IPN Color palette scale: color}
\usage{
scale_color_ipn(fl = 5, name, labels)
}
\arguments{
\item{fl}{set Forschungslinie, default set to "6" which is IPN blue}

\item{name}{set scale name}

\item{labels}{set scale labels}
}
\description{
IPN Color palette scale: color
}
\examples{
scale_color_ipn( fl = 2 name = c("example scale"), labels = c("A","B"))

}
