% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipn_scale_basic.R
\name{scale_fill_ipn}
\alias{scale_fill_ipn}
\title{IPN Color palette scale: fill}
\usage{
scale_fill_ipn(fl = 6, name, labels)
}
\arguments{
\item{fl}{set Forschungslinie, default set to "6" which is IPN blue}

\item{name}{set sclae name}

\item{labels}{set scale labels}
}
\description{
IPN Color palette scale: fill
}
\examples{
scale_fill_ipn( fl = 2 name = c("example scale"), labels = c("A","B"))

}
